requirements.txt
pbr>=0.5.21,<1.0
Babel>=0.9.6
SQLAlchemy>=0.7.8,<0.7.99
oslo.config>=1.2.0
iso8601>=0.1.4
